[chore: refactor] moved around component files to relevant packages
