+++ app/src/main/java/dev/logickoder/qrpay/ui/screens/home/BalanceSummaryCard.kt
@@ -1,4 +1,4 @@
-package dev.logickoder.qrpay.ui.shared.component
+package dev.logickoder.qrpay.ui.screens.home
 
 import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
@@ -10,6 +10,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.constraintlayout.compose.ConstraintLayout
 import dev.logickoder.qrpay.R
+import dev.logickoder.qrpay.ui.shared.composables.Icon
+import dev.logickoder.qrpay.ui.shared.modifiers.nonExistent
 import dev.logickoder.qrpay.ui.theme.Theme
 import dev.logickoder.qrpay.ui.theme.TintedText
 import dev.logickoder.qrpay.utils.Amount
