+++ app/src/main/java/dev/logickoder/qrpay/ui/shared/composables/Buttons.kt
@@ -1,4 +1,4 @@
-package dev.logickoder.qrpay.ui.shared.component
+package dev.logickoder.qrpay.ui.shared.composables
 
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonDefaults
