+++ app/src/main/java/dev/logickoder/qrpay/ui/shared/composables/RadioBox.kt
@@ -1,4 +1,4 @@
-package dev.logickoder.qrpay.ui.shared.component
+package dev.logickoder.qrpay.ui.shared.composables
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
